requirements.txt
flask==2.3.3
flask-socketio==5.3.4
pyserial==3.5
opencv-python==4.8.0.76
numpy==1.24.3
paho-mqtt==2.2.1
python-dotenv==1.0.0
RPi.GPIO==0.7.1
picamera2==0.3.12 